fix(cache): register server method on server creation
